@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,26 +29,28 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  google_fonts: ^5.1.0
-  responsive_framework: ^0.1.5
-  font_awesome_flutter: ^9.0.0-nullsafety
-  smooth_page_indicator: ^1.1.0
-  carousel_slider: ^4.0.0-nullsafety.0
+  google_fonts: ^6.1.0
+  responsive_framework: ^1.1.1
+  font_awesome_flutter: ^10.7.0
+  smooth_page_indicator: ^1.2.0
+  flutter_carousel_widget: ^3.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  url_launcher: ^6.1.0
-  http: ^1.0.0
+  url_launcher: ^6.3.1
+  http: ^1.2.2
   
 
-  flutter_svg: ^1.1.6
+  flutter_svg: ^2.0.10+1
   graphql_flutter: ^5.2.0-beta.5
   super_tooltip: ^2.0.5
   logger: ^2.0.1
-  photo_view: ^0.14.0
+  photo_view: ^0.15.0
   
-  
+  # State Management
+  flutter_riverpod: ^2.6.1
+  riverpod_annotation: ^2.3.5
 
 dev_dependencies:
   flutter_test:
@@ -60,7 +62,11 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^4.0.0
+  
+  # Code generation for Riverpod
+  riverpod_generator: ^2.4.3
+  build_runner: ^2.4.13
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
