@@ -6,9 +6,9 @@ import '../../../../constants/constants.dart';
 class BorderedText extends StatelessWidget {
   final String text;
   const BorderedText({
-    Key? key,
+    super.key,
     required this.text,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
